orderPage table

+++ src/action/productAction.js
@@ -1,6 +1,5 @@
 import api from "../utils/api";
 import * as types from "../constants/product.constants";
-import { toast } from "react-toastify";
 import { commonUiActions } from "./commonUiAction";
 
 /** Redux Thunk
@@ -8,22 +7,51 @@ import { commonUiActions } from "./commonUiAction";
  * 따라서 redux thunk 미들웨어를 사용하여 비동기적 처리를 한다.
  * https://redux.js.org/usage/writing-logic-thunks
  */
-const getProductList = (query) => async (dispatch) => {
+
+//MainPage
+const getProductList = (query, navigate) => async (dispatch) => {
   try {
     dispatch({ type: types.PRODUCT_GET_REQUEST });
 
     const res = await api.get("/product", { params: { ...query } });
-
+    console.log(res);
     if (res.status !== 200) {
       throw new Error(res.error);
     }
 
     dispatch({ type: types.PRODUCT_GET_SUCCESS, payload: res.data });
   } catch (err) {
-    dispatch({ type: types.PRODUCT_GET_FAIL, payload: err.error });
+    dispatch({
+      type: types.PRODUCT_GET_FAIL,
+      payload: err.error,
+    });
+    dispatch(commonUiActions.showToastMessage(err.error, "error"));
+    navigate("/");
   }
 };
 
+//AdminPage
+const getAdminProductList = (query, navigate) => async (dispatch) => {
+  try {
+    dispatch({ type: types.PRODUCT_GET_REQUEST });
+
+    const res = await api.get("/product/admin", { params: { ...query } });
+    console.log(res);
+    if (res.status !== 200) {
+      throw new Error(res.error);
+    }
+
+    dispatch({ type: types.PRODUCT_GET_SUCCESS, payload: res.data });
+  } catch (err) {
+    dispatch({
+      type: types.PRODUCT_GET_FAIL,
+      payload: err.error,
+    });
+    dispatch(commonUiActions.showToastMessage(err.error, "error"));
+    navigate("/");
+  }
+};
+//DetailPage
 const getProductDetail = (id) => async (dispatch) => {
   try {
     dispatch({ type: types.GET_PRODUCT_DETAIL_REQUEST });
@@ -42,6 +70,7 @@ const getProductDetail = (id) => async (dispatch) => {
   }
 };
 
+//AdminPage
 const createProduct = (formData) => async (dispatch) => {
   try {
     dispatch({ type: types.PRODUCT_CREATE_REQUEST });
@@ -62,7 +91,7 @@ const deleteProduct = (id) => async (dispatch) => {
   /**delete api 호출 */
   try {
     dispatch({ type: types.PRODUCT_DELETE_REQUEST });
-    const res = await api.delete(`product/${id}`);
+    const res = await api.delete(`product/admin/${id}`);
     if (res.status !== 200) {
       throw new Error(res.error);
     }
@@ -79,7 +108,7 @@ const deleteProduct = (id) => async (dispatch) => {
 const editProduct = (formData, id) => async (dispatch) => {
   try {
     dispatch({ type: types.PRODUCT_EDIT_REQUEST });
-    const res = await api.put(`/product/${id}`, formData);
+    const res = await api.put(`/product/admin/${id}`, formData);
     if (res.status !== 200) {
       throw new Error(res.error);
     }
@@ -98,4 +127,5 @@ export const productActions = {
   deleteProduct,
   editProduct,
   getProductDetail,
+  getAdminProductList,
 };
